# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpsMind Core API","description":"Phase 1 foundation: tenant, auth, alert ingestion, incident lifecycle","version":"phase1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/api/projects":{"get":{"tags":["tenant-controller"],"operationId":"listProjects","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProjectResponse"}}}}}}},"post":{"tags":["tenant-controller"],"operationId":"createProject","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProjectResponse"}}}}}}},"/api/projects/{projectId}/services":{"get":{"tags":["tenant-controller"],"operationId":"listServices","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ServiceResponse"}}}}}}},"post":{"tags":["tenant-controller"],"operationId":"createService","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateServiceRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceResponse"}}}}}}},"/api/projects/{projectId}/environments":{"get":{"tags":["tenant-controller"],"operationId":"listEnvironments","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EnvironmentResponse"}}}}}}},"post":{"tags":["tenant-controller"],"operationId":"createEnvironment","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEnvironmentRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EnvironmentResponse"}}}}}}},"/api/investigations/{investigationId}/run":{"post":{"tags":["investigation-controller"],"operationId":"rerun","parameters":[{"name":"investigationId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InvestigationResponse"}}}}}}},"/api/incidents/{incidentId}/investigations":{"get":{"tags":["investigation-controller"],"operationId":"listForIncident","parameters":[{"name":"incidentId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InvestigationResponse"}}}}}}},"post":{"tags":["investigation-controller"],"operationId":"create","parameters":[{"name":"incidentId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InvestigationResponse"}}}}}}},"/api/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterOrganizationRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/alerts":{"get":{"tags":["alert-controller"],"operationId":"list_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Alert"}}}}}}},"post":{"tags":["alert-controller"],"operationId":"ingest","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlertIngestRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AlertResponse"}}}}}}},"/api/alerts/webhooks/alertmanager":{"post":{"tags":["alertmanager-webhook-controller"],"operationId":"receive","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlertmanagerWebhookPayload"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/incidents/{id}/status":{"patch":{"tags":["incident-controller"],"operationId":"updateStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IncidentStatusUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IncidentResponse"}}}}}}},"/api/investigations/{investigationId}":{"get":{"tags":["investigation-controller"],"operationId":"get","parameters":[{"name":"investigationId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InvestigationResponse"}}}}}}},"/api/incidents":{"get":{"tags":["incident-controller"],"operationId":"list","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/IncidentResponse"}}}}}}}},"/api/incidents/{id}":{"get":{"tags":["incident-controller"],"operationId":"get_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IncidentResponse"}}}}}}},"/api/incidents/{id}/timeline":{"get":{"tags":["incident-controller"],"operationId":"timeline","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/IncidentTimelineEntryResponse"}}}}}}}}},"components":{"schemas":{"CreateProjectRequest":{"required":["name","slug"],"type":"object","properties":{"name":{"type":"string"},"slug":{"pattern":"^[a-z0-9-]+$","type":"string"}}},"ProjectResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"organizationId":{"type":"string","format":"uuid"},"name":{"type":"string"},"slug":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"CreateServiceRequest":{"required":["name","slug"],"type":"object","properties":{"name":{"type":"string"},"slug":{"pattern":"^[a-z0-9-]+$","type":"string"},"description":{"type":"string"}}},"ServiceResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"projectId":{"type":"string","format":"uuid"},"name":{"type":"string"},"slug":{"type":"string"},"description":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"CreateEnvironmentRequest":{"required":["type"],"type":"object","properties":{"type":{"type":"string","enum":["LOCAL","STAGING","PRODUCTION"]}}},"EnvironmentResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"projectId":{"type":"string","format":"uuid"},"type":{"type":"string","enum":["LOCAL","STAGING","PRODUCTION"]},"createdAt":{"type":"string","format":"date-time"}}},"EvidenceResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"type":{"type":"string","enum":["METRIC","LOG","TRACE","DEPLOYMENT","CONFIG_CHANGE","DEPENDENCY","HISTORICAL_INCIDENT"]},"source":{"type":"string"},"observedAt":{"type":"string","format":"date-time"},"title":{"type":"string"},"description":{"type":"string"},"observedValue":{"type":"string"}}},"InvestigationResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"incidentId":{"type":"string","format":"uuid"},"status":{"type":"string","enum":["PENDING","RUNNING","COMPLETED","FAILED"]},"failureReason":{"type":"string"},"aiProvider":{"type":"string"},"aiModel":{"type":"string"},"startedAt":{"type":"string","format":"date-time"},"completedAt":{"type":"string","format":"date-time"},"evidence":{"type":"array","items":{"$ref":"#/components/schemas/EvidenceResponse"}},"rcaFinding":{"$ref":"#/components/schemas/RcaFindingResponse"}}},"RcaFindingResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"rootCause":{"type":"string"},"hypothesis":{"type":"string"},"confidenceLevel":{"type":"string"},"confidenceScore":{"type":"number","format":"double"},"affectedService":{"type":"string"},"impact":{"type":"string"},"supportingEvidenceIds":{"type":"array","items":{"type":"string","format":"uuid"}},"reasoningSummary":{"type":"string"},"recommendedActions":{"type":"array","items":{"type":"string"}},"modelProvider":{"type":"string"},"modelName":{"type":"string"},"generatedAt":{"type":"string","format":"date-time"}}},"RegisterOrganizationRequest":{"required":["email","fullName","organizationName","organizationSlug","password"],"type":"object","properties":{"organizationName":{"type":"string"},"organizationSlug":{"pattern":"^[a-z0-9-]+$","type":"string"},"fullName":{"type":"string"},"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"}}},"AuthResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"expiresInSeconds":{"type":"integer","format":"int64"},"userId":{"type":"string","format":"uuid"},"organizationId":{"type":"string","format":"uuid"},"role":{"type":"string"}}},"LoginRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"AlertIngestRequest":{"required":["alertType","environmentId","projectId","severity","source","title"],"type":"object","properties":{"projectId":{"type":"string","format":"uuid"},"environmentId":{"type":"string","format":"uuid"},"serviceId":{"type":"string","format":"uuid"},"source":{"type":"string"},"alertType":{"type":"string"},"severity":{"type":"string","enum":["LOW","MEDIUM","HIGH","CRITICAL"]},"title":{"type":"string"},"description":{"type":"string"},"metricName":{"type":"string"},"thresholdValue":{"type":"number"},"currentValue":{"type":"number"},"timestamp":{"type":"string","format":"date-time"},"metadata":{"type":"object","additionalProperties":{"type":"object"}}}},"AlertResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"projectId":{"type":"string","format":"uuid"},"environmentId":{"type":"string","format":"uuid"},"serviceId":{"type":"string","format":"uuid"},"source":{"type":"string"},"alertType":{"type":"string"},"severity":{"type":"string","enum":["LOW","MEDIUM","HIGH","CRITICAL"]},"status":{"type":"string","enum":["RECEIVED","LINKED"]},"title":{"type":"string"},"description":{"type":"string"},"metricName":{"type":"string"},"thresholdValue":{"type":"number"},"currentValue":{"type":"number"},"receivedAt":{"type":"string","format":"date-time"},"createdIncidentId":{"type":"string","format":"uuid"},"createdIncidentNumber":{"type":"string"}}},"AlertmanagerAlert":{"type":"object","properties":{"status":{"type":"string"},"labels":{"type":"object","additionalProperties":{"type":"string"}},"annotations":{"type":"object","additionalProperties":{"type":"string"}},"startsAt":{"type":"string"},"endsAt":{"type":"string"},"generatorURL":{"type":"string"},"fingerprint":{"type":"string"}}},"AlertmanagerWebhookPayload":{"type":"object","properties":{"version":{"type":"string"},"groupKey":{"type":"string"},"status":{"type":"string"},"receiver":{"type":"string"},"groupLabels":{"type":"object","additionalProperties":{"type":"string"}},"commonLabels":{"type":"object","additionalProperties":{"type":"string"}},"commonAnnotations":{"type":"object","additionalProperties":{"type":"string"}},"externalURL":{"type":"string"},"alerts":{"type":"array","items":{"$ref":"#/components/schemas/AlertmanagerAlert"}}}},"IncidentStatusUpdateRequest":{"required":["status"],"type":"object","properties":{"status":{"type":"string","enum":["DETECTED","NEW","INVESTIGATING","MITIGATING","AWAITING_APPROVAL","REMEDIATING","VERIFYING","RESOLVED","ESCALATED","CLOSED"]},"note":{"type":"string"}}},"IncidentResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"incidentNumber":{"type":"string"},"projectId":{"type":"string","format":"uuid"},"environmentId":{"type":"string","format":"uuid"},"primaryServiceId":{"type":"string","format":"uuid"},"title":{"type":"string"},"severity":{"type":"string","enum":["LOW","MEDIUM","HIGH","CRITICAL"]},"status":{"type":"string","enum":["DETECTED","NEW","INVESTIGATING","MITIGATING","AWAITING_APPROVAL","REMEDIATING","VERIFYING","RESOLVED","ESCALATED","CLOSED"]},"summary":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"resolvedAt":{"type":"string","format":"date-time"},"closedAt":{"type":"string","format":"date-time"},"allowedNextStatuses":{"uniqueItems":true,"type":"array","items":{"type":"string","enum":["DETECTED","NEW","INVESTIGATING","MITIGATING","AWAITING_APPROVAL","REMEDIATING","VERIFYING","RESOLVED","ESCALATED","CLOSED"]}}}},"IncidentTimelineEntryResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"eventType":{"type":"string","enum":["INCIDENT_CREATED","STATUS_CHANGED","ALERT_LINKED","NOTE_ADDED"]},"description":{"type":"string"},"actor":{"type":"string"},"occurredAt":{"type":"string","format":"date-time"}}},"Alert":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"organizationId":{"type":"string","format":"uuid"},"projectId":{"type":"string","format":"uuid"},"environmentId":{"type":"string","format":"uuid"},"serviceId":{"type":"string","format":"uuid"},"source":{"type":"string"},"alertType":{"type":"string"},"severity":{"type":"string","enum":["LOW","MEDIUM","HIGH","CRITICAL"]},"status":{"type":"string","enum":["RECEIVED","LINKED"]},"title":{"type":"string"},"description":{"type":"string"},"metricName":{"type":"string"},"thresholdValue":{"type":"number"},"currentValue":{"type":"number"},"rawPayload":{"type":"string"},"receivedAt":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"}}}},"securitySchemes":{"bearerAuth":{"type":"http","name":"bearerAuth","scheme":"bearer","bearerFormat":"JWT"}}}}